requirements.txt
streamlit
matplotlib
seaborn
plotly
pandas
statsmodels
scikit-learn
Pillow
numpy
requests
